Reduced MTU to 1300 to support NordVPN.
See conversation starting here on Discord - https://discord.com/channels/449263083769036810/1067070048830771200/1092846833207869531

If that's gone in future, the TLDR; is that there's no one exact number that works across the board here, and 1400 is clearly too high for NordVPN. I was able to get it to work at 1390 but decided to reduce it to 1300 to give a bit of "buffer". Lower it is, the more supported it is but the worse bandwidth use can be.