After a lengthy investigation with Unity on Discord, this is the fix for the issue with us running out of handles.
The problem is that there was a buffer overrun which caused a Lock variable to be trashed/stomped-over. This then resulted in a section of code failing to obtain a lock and then, in that event, also failing to release the handle - so it leaked.

This fix addresses the failure to release the handle in the event of this error - but, more crucially, it fixes the buffer overrun that caused it in the first place.

NOTE: The final fix came from Simon @ Unity via Discord. They will be released a hotfix that we'll then want to merge, but this will be the contents of that hotfix.